[fix]: delete app test case
